Type:fix
Descripton:When the expression is wrong, input the number and equal then
re-enter a new expression, the error expression in the history record
has not been replaced.
An expression error reenters a new expression with no result in the input edit.

RootCause:Enter the wrong judgment.

Solution:Reset the judgement.